+++ plugin/buildSrc/src/main/kotlin/PluginDependencies.kt
@@ -3,7 +3,6 @@ object PluginVersions {
     val ktlintPlugin = "6.2.1"
     val gradlePublishPlugin = "0.10.0"
     val androidPlugin = "3.2.0"
-    val kotlinNativePlugin = "0.8.2"
     val semver = "1.1.0"
     val gradleWrapper = "4.10.2"
     val junit = "4.12"